extract OntologyURI to fix circular imports
